+++ src/GitNoob.Gui.Forms/ActionButton.cs
@@ -7,12 +7,14 @@ namespace GitNoob.Gui.Forms
     public class ActionButton : Button
     {
         private Gui.Program.Action.IAction _action;
+        private ContextMenu _context;
 
-        public ActionButton(ToolTip toolTips, string text, Gui.Program.Action.IAction action, ref Point location) : base()
+        public ActionButton(ToolTip toolTips, string text, ContextMenu context, Gui.Program.Action.IAction action, ref Point location) : base()
         {
             const int margin = 8;
 
             _action = action;
+            _context = context;
 
             this.ClientSize = new Size(48, 48);
             this.AutoSize = true;
@@ -30,6 +32,7 @@ namespace GitNoob.Gui.Forms
                 }
 
                 this.Click += ActionButton_Click;
+                this.MouseDown += ActionButton_MouseDown;
             }
 
             this.Location = location;
@@ -39,6 +42,19 @@ namespace GitNoob.Gui.Forms
         private void ActionButton_Click(object sender, EventArgs e)
         {
             _action.execute();
+        }
+
+        private void ActionButton_MouseDown(object sender, MouseEventArgs e)
+        {
+            switch (e.Button)
+            {
+                case MouseButtons.Right:
+                    if (_context != null)
+                    {
+                        _context.Show(this, new Point(0, this.Height));
+                    }
+                    break;
+            }
         }
     }
 }